user and role management improvements

+++ Oqtane.Server/Controllers/UserRoleController.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authorization;
 using Oqtane.Enums;
@@ -6,6 +6,7 @@ using Oqtane.Models;
 using Oqtane.Shared;
 using Oqtane.Infrastructure;
 using Oqtane.Repository;
+using System.Linq;
 
 namespace Oqtane.Controllers
 {
@@ -13,13 +14,15 @@ namespace Oqtane.Controllers
     public class UserRoleController : Controller
     {
         private readonly IUserRoleRepository _userRoles;
+        private readonly IRoleRepository _roles;
         private readonly ITenantResolver _tenants;
         private readonly ISyncManager _syncManager;
         private readonly ILogManager _logger;
 
-        public UserRoleController(IUserRoleRepository userRoles, ITenantResolver tenants, ISyncManager syncManager, ILogManager logger)
+        public UserRoleController(IUserRoleRepository userRoles, IRoleRepository roles, ITenantResolver tenants, ISyncManager syncManager, ILogManager logger)
         {
             _userRoles = userRoles;
+            _roles = roles;
             _syncManager = syncManager;
             _tenants = tenants;
             _logger = logger;
@@ -46,8 +49,13 @@ namespace Oqtane.Controllers
         [Authorize(Roles = RoleNames.Admin)]
         public UserRole Post([FromBody] UserRole userRole)
         {
-            if (ModelState.IsValid)
+            var role = _roles.GetRole(userRole.RoleId);
+            if (ModelState.IsValid && (User.IsInRole(RoleNames.Host) || role.Name != RoleNames.Host))
             {
+                if (role.Name == RoleNames.Host)
+                {
+                    _userRoles.DeleteUserRoles(userRole.UserId);
+                }
                 userRole = _userRoles.AddUserRole(userRole);
                 _syncManager.AddSyncEvent(_tenants.GetTenant().TenantId, EntityNames.User, userRole.UserId);
                 _logger.Log(LogLevel.Information, this, LogFunction.Create, "User Role Added {UserRole}", userRole);
@@ -60,7 +68,8 @@ namespace Oqtane.Controllers
         [Authorize(Roles = RoleNames.Admin)]
         public UserRole Put(int id, [FromBody] UserRole userRole)
         {
-            if (ModelState.IsValid)
+            var role = _roles.GetRole(userRole.RoleId);
+            if (ModelState.IsValid && (User.IsInRole(RoleNames.Host) || role.Name != RoleNames.Host))
             {
                 userRole = _userRoles.UpdateUserRole(userRole);
                 _syncManager.AddSyncEvent(_tenants.GetTenant().TenantId, EntityNames.User, userRole.UserId);
@@ -75,9 +84,17 @@ namespace Oqtane.Controllers
         public void Delete(int id)
         {
             UserRole userRole = _userRoles.GetUserRole(id);
-            _userRoles.DeleteUserRole(id);
-            _syncManager.AddSyncEvent(_tenants.GetTenant().TenantId, EntityNames.User, userRole.UserId);
-            _logger.Log(LogLevel.Information, this, LogFunction.Delete, "User Role Deleted {UserRole}", userRole);
+            if (User.IsInRole(RoleNames.Host) || userRole.Role.Name != RoleNames.Host)
+            {
+                _userRoles.DeleteUserRole(id);
+                if (userRole.Role.Name == RoleNames.Host)
+                {
+                    var role = _roles.GetRoles(_tenants.GetAlias().SiteId).FirstOrDefault(item => item.Name == RoleNames.Registered);
+                    _userRoles.AddUserRole(new UserRole { UserId = userRole.UserId, RoleId = role.RoleId, EffectiveDate = null, ExpiryDate = null });
+                }
+                _syncManager.AddSyncEvent(_tenants.GetTenant().TenantId, EntityNames.User, userRole.UserId);
+                _logger.Log(LogLevel.Information, this, LogFunction.Delete, "User Role Deleted {UserRole}", userRole);
+            }
         }
     }
 }
